requirements.txt
streamlit
pytesseract
opencv-python
pillow
python-docx